requirements.txt
click==8.1.8
pydantic==2.6.3
pyobjc==9.2; platform_system=="Darwin"
python-frontmatter==1.0.0
qdrant-client==1.12.0
rich==13.7.0
sentence-transformers==3.0.1
torch>=1.11.0
huggingface-hub>=0.23.0
sqlalchemy==2.0.40
watchdog==6.0.0 